Split src/lib and src/intl
