+++ src/intl/resolver.js
